[MOVE] snDetail to usused
